fix: Windows Arduino IDE 스케치 폴더 밖 include 오류 해결 (로컬 전용, 미푸시)
- 원인: EthernetClient등과 무관하게, "#include ../../shared/protocol_version.h"
  방식이 Linux arduino-cli에서는 컴파일됐지만 Windows Arduino IDE로 스케치를
  단독으로 열어 컴파일하면 fatal error(파일 없음)로 실패. Arduino IDE가 스케치
  빌드 시 스케치 폴더 밖(../../)의 상대경로 include를 안정적으로 못 찾는 구조적
  문제(파일이 실제 존재하고 경로가 맞아도 실패).
- 수정: shared/protocol_version.h를 기준 원장(canonical)으로 유지하되,
  firmware/arm_relay_6ch/protocol_version.h, firmware/main_eth_8di_8ro/
  protocol_version.h에 byte-for-byte 동일한 사본을 생성. 모든 include를
  스케치 폴더 내부만 가리키는 #include "protocol_version.h"로 통일(8개 파일
  + Windows 실제 config.h 2개, 총 10곳) — Arduino의 표준 tab-include 방식이라
  플랫폼/IDE 무관하게 안정적으로 동작
- scripts/sync_protocol_version.sh 신규: 옵션 없이 실행하면 기준원장→사본
  동기화, --check로 세 파일 byte-for-byte 일치 여부 검사(불일치 시 실패
  종료코드) — 수동 파일 복사가 아니라 재현 가능한 스크립트로 관리
- shared/protocol_version.h 최상단에 "기준원장 위치 + 사본 직접수정 금지"
  배너를 넣어 세 파일 어느 것을 열어도 원장 위치를 알 수 있게 함
- PROTOCOL_VERSION 매크로 이름/값은 세 파일 모두 동일하므로 rs485_master.cpp의
  기존 런타임 버전 비교 로직(isFlowSupported, 매 RS485 폴링마다 비교)은 변경
  없이 그대로 유지됨
- 절대경로(C:\Users\...)는 코드에 넣지 않음, config.h의 실제 MQTT 비밀정보는
  include 한 줄만 고치고 손대지 않음
- 부가 수정: mqtt_manager.cpp의 %X/uint32_t 포맷 타입 불일치 경고(이번 재컴파일
  중 --warnings all로 발견, 본 작업과 무관하지만 안전한 1줄 캐스팅으로 정리)
- Linux arduino-cli로 재컴파일 완료(양쪽 성공, 우리 코드 자체 경고 0건 —
  라이브러리 내부(eModbus/SSLClient) 경고는 별개, 우리 코드 아님). **Windows
  Arduino IDE에서의 실제 재현/검증은 이 환경에서 수행할 수 없어 미완료 —
  사용자가 직접 재시도 필요**

+++ smartfarm-wired-controller/firmware/main_eth_8di_8ro/mqtt_manager.cpp
@@ -13,8 +13,11 @@ void MqttManager::begin(const char* host, uint16_t port, const char* user, const
   // 만든다 — 기존 WiFi ctlr-0004와 새 유선 메인 노드가 전환/롤백 테스트 중 동시에
   // 켜져 있어도 같은 clientId 때문에 브로커가 한쪽을 끊어버리는 사고를 방지한다.
   uint64_t mac = ESP.getEfuseMac();
+  // %X는 unsigned int를 기대하는데 uint32_t는 이 플랫폼에서 "long unsigned int"라
+  // 타입이 달라 -Wformat 경고가 남는다(값 자체는 동일 크기라 문제 없었지만 2026-08-23
+  // --warnings all 재검증 중 발견해 명시적으로 unsigned int로 캐스팅해 정리).
   snprintf(clientId_, sizeof(clientId_), "%s-%04X%08X", clientIdPrefix,
-           (uint16_t)(mac >> 32), (uint32_t)mac);
+           (unsigned int)(uint16_t)(mac >> 32), (unsigned int)(uint32_t)mac);
   Serial.printf("[MQTT] clientId: %s\n", clientId_);
 
   // 순수 호스트명만 허용 — 프로토콜/포트가 섞여 들어오는 흔한 실수 방어
